requirements.txt
uagents
openai
python-dotenv